@ralfbiedert 2023-08-26
Thanks for the PR. I'd be willing to (temporarily) link a lightweight 3rd party library for faster YUV - RGB conversion, although it would be nicer to just optimize our code. 

However, the crate you propose is another native lib with complex build logic that downloads 3rd party code during compilation. 

While compiling the native OpenH264 C lib to this crate was a necessity, I very much want to avoid other native libs, or anything that complicates building this crate any further.