@soheilsalimidev 2023-08-27
I agree this lib is complex and in most scenarios unnecessary but I think it is a good feature to have if someone is willing to pay the cost of having another c lib or at least can be added as an example

PS: I searched for a pure rust alternative to this lib or anything fast enough(I use this for streaming video) for this conversion but I had no luck finding one

anyway if you think it's not necessary then I think you should close it. Thanks for reviewing the PR