@ralfbiedert 2023-08-28
Alright, closing for now. I think it's better for people to link this crate themselves if they have a need for speed. That said, PRs for other format conversions in our lib are welcome.